refactor: standardise file naming to kebab case
Update import paths for renamed files
