+++ src/main.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { ChakraProvider } from '@chakra-ui/react';
 import ReactDOM from 'react-dom/client';
 
-import AppRouter from './AppRouter';
+import AppRouter from './router';
 import { system } from './theme';
 
 ReactDOM.createRoot(document.getElementById('root')).render(
